Change the namespace of the general services classes
